This is the rest of the Admins page for functionality. The ability to add new Fund Types has been made available under the administration tab. There is also the ability to edit each active fund type. Active meaning any that are not expired. In order to get everything working correctly i had to make one adjustment to the database file so the fund_type table column ID would auto increment to work with the hibernate calls.
When editing the fund type the dates were returning as long sets of numbers so i have made them blank for now. Conversion will be needed to display correctly.

Also need to use angular to modify the header.jsp to order to only show admin button when on an admin account.